Combine main page and left side menu
