+++ frontend/src/app/app-routing.module.ts
@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { MainPageComponent } from './components/main-page/main-page.component';
 
 const routes: Routes = [
-  { path: '', component: MainPageComponent },
   {
     path: '',
     loadChildren: () => import('./moduls/auth/auth.module').then((m) => m.AuthModule)
