--- frontend/src/app/components/main-page/main-page.component.html
@@ -1,135 +1,0 @@
-<main> 
-    <div class="accordion">
-        <!-- Navigation arrows -->
-        <div class="arrows">
-            <button (click)="nextSlide()" class="arrow-left">
-                <i class="chevron left icon"></i>
-            </button>
-            <button (click)="previousSlide()" class="arrow-right">
-                <i class="chevron right icon"></i>
-            </button>
-        </div>
-        <!-- Profile icon in top right corner -->
-        <button class="profile" (click)="showProfileMenu()">
-            <img class="ui avatar image" src="https://s3-alpha-sig.figma.com/img/5ba6/3340/8981907adb4cdc9c14be2ac429a30ff6?Expires=1628467200&Signature=F0KRMuqseJflr9PZvKCf6GeWuVkI9hxAeXrvKSVMPdGcHY98Sltas2~pAn~wNyFkRV3Inyn-qL1f5-k~HLx58EeYG75xZSwBIpWNZVOUp8vrB2jKl9ojIT557f7OPVQ0yYGx6vjTCGB8uPbaGp9DWR9JJVuzpuyR0kn~eRTmaO8cveSQDC9rDf-rizQ3JeupdUS-KBv-InGcQ5u7tQLjD1QsUFXLtLg7AB-ZNxPpy6xB9bZKK44HUOu8G6NZIpZOY6cLILsfKwGx-ULoGo8jubwtNrspmAf8N0Zi-JrQmuq1c0JeNdankoH8Mp0CR0b~N~7~voG7LGo16pKRTjJHMg__&Key-Pair-Id=APKAINTVSUGEWH5XD5UA">
-            <span>marc_dorwart</span><i class="caret down icon"></i>
-            <div id="profile-menu" class="dropdown-content">
-                <a href="#"><i class="user icon"></i> My Profile</a>
-                <a href="#"><i class="setting icon"></i> Settings</a>
-                <a href="#"><i class="external alternate icon"></i> Perflow studio</a>
-                <a href="#"><i class="sign-out icon"></i> Log Out</a>
-            </div>
-        </button>
-        <!-- Accordion Information -->
-        <div class="accordion-information">
-            <img class="album-icon" alt="Album" src="https://s3-alpha-sig.figma.com/img/d75c/1009/f21231d6f5d9480e4025fca6081aa434?Expires=1628467200&Signature=OTjTRmC-15s41yTDgmE8aKF5q6CnBmJAo55dQRIG7K0XOj8A9L3ENDn6dN4EmFOukNiA7DkyXDUXbf3bAjngVFLV7yUgxTssDto6jzNDJGShiDViBnpZAvV6BSpHRuQYIkGj4vTtKJZLeNtDQXQgnMrNLLz8LA1m1XHxLhvug7dLrmXAO2hxHhJk4PUqXplBNEnxyPSZjs-sY44I--97FKwspfh1eyjIlEIdj9eowU4WNh52N-mrnShdCdKJpbaIzE5-mJkpug91XKth-IWB5iHnotKb88bARs1rCFiMaZlp5-Xxi6jpJdOeYIriyilk9~obp6ItPZK-3t2OxS7tag__&Key-Pair-Id=APKAINTVSUGEWH5XD5UA">
-            <div class="album-info">
-                <p class="type">Album</p>
-                <p class="album-title">Perfect Morning</p>
-                <p class="album-description">Lorem ipsum dolor sit amet consectetur adipisicing elit. Maiores facilis deleniti quidem atque obcaecati eaque quod tenetur odit laborum ex explicabo iste placeat, repellendus aliquid alias voluptatibus, quos incidunt? Quos.</p>
-                <div class="buttons">
-                    <a href="#" class="btn-play">Play</a>
-                    <a href="#" class="btn-save">Save</a>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="ui grid">
-        <!-- 2. Recently played -->
-        <div class="ui sixteen column row recently-played">
-            <div class="header">Recently played</div>
-            <div class="ui grid recently-played-cont">
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-                <app-recently-played-card></app-recently-played-card>
-            </div>
-        </div>
-
-
-        <!-- 3. New Releases -->
-        <div class="ui sixteen column row new-releases">
-            <div class="header">New Releases</div>
-            <div id="newReleases" class="new-releases-cont">
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-                <app-new-releases-card></app-new-releases-card>
-            </div>
-            <div class="scroll-box">
-                <button class="scroll-button" (click)="scroll('newReleases')">
-                    <i class="chevron right icon"></i>
-                </button>
-            </div>
-        </div>
-    
-        
-        <!-- 4. Calm rhythms -->
-        <div class="sixteen column row calm-rhythms">  
-            <div class="header">Calm rhythms</div>
-            <div id="calmRhythms" class="calm-rhythms-cont">
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-                <app-calm-rhythms-card></app-calm-rhythms-card>
-            </div>
-            <div class="scroll-box">
-                <button class="scroll-button" (click)="scroll('calmRhythms')">
-                    <i class="chevron right icon"></i>
-                </button>
-            </div>
-        </div>
-    
-
-        <!-- 5. Your mix -->
-        <div class="sixteen column row your-mix">
-            <div class="header">Your mix</div>
-            <div id="yourMix"  class="ui grid your-mix-cont">
-                <app-your-mix-card></app-your-mix-card>
-                <app-your-mix-card></app-your-mix-card>
-                <app-your-mix-card></app-your-mix-card>
-                <app-your-mix-card></app-your-mix-card>
-                <app-your-mix-card></app-your-mix-card>
-                <app-your-mix-card></app-your-mix-card>
-            </div>
-        </div>
-
-    
-        <!-- 6. Top 100 songs -->
-        <div class="sixteen column row calm-rhythms">
-            <div class="header">Top 100 songs</div>
-            <div id="top100Songs" class="calm-rhythms-cont">
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-               <app-top-songs-card></app-top-songs-card>
-            </div>
-            <div class="scroll-box">
-                <button class="scroll-button" (click)="scroll('top100Songs')">
-                    <i class="chevron right icon"></i>
-                </button>
-            </div> 
-        </div>       
-    </div>
-</main>
