+++ frontend/src/app/moduls/main/cards/calm-rhythms-card/calm-rhythms-card.component.sass
@@ -42,6 +42,7 @@
         font-size: 16px
         line-height: 19px
         text-align: left
+        text-overflow: ellipsis
         color: #FAFBFF
         margin: 0px 16px 5px 16px 
     .card-desc
@@ -52,14 +53,9 @@
         font-size: 14px
         line-height: 17px
         text-align: left
+        text-overflow: ellipsis
         color: #FAFBFF
         margin: 0px 16px 17px 16px
 
-
-
-
-
-.border
-    border: 1px solid red
