+++ frontend/src/app/moduls/main/cards/new-releases-card/new-releases-card.component.sass
@@ -1,7 +1,7 @@
 .nr-card
     height: 270px
     width: 214px
-    margin-right: 46px 
+    margin-right: 46px
     .fader
         text-align: center
         position: relative
@@ -38,6 +38,7 @@
         font-weight: 500
         font-size: 16px
         line-height: 19px
+        text-overflow: ellipsis
         text-align: center
         color: #FAFBFF
     .card-desc
@@ -47,14 +48,6 @@
         font-weight: 300
         font-size: 14px
         line-height: 17px
+        text-overflow: ellipsis
         text-align: center
         color: #FAFBFF
-
-
-
-
-
-.border
-    border: 1px solid red
-
-
