+++ frontend/src/app/moduls/main/cards/new-releases-card/new-releases-card.component.spec.ts
@@ -8,9 +8,9 @@ describe('NewReleasesCardComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ NewReleasesCardComponent ]
+      declarations: [NewReleasesCardComponent]
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
