+++ frontend/src/app/moduls/main/cards/new-releases-card/new-releases-card.component.ts
