+++ frontend/src/app/moduls/main/cards/recently-played-card/recently-played-card.component.sass
@@ -54,6 +54,7 @@
         font-weight: normal
         font-size: 14px
         line-height: 17px
+        text-overflow: ellipsis
         color: #FAFBFF
         margin: 9px 0px 0px 15px
         cursor: pointer
@@ -65,12 +66,8 @@
         font-weight: 300
         font-size: 14px
         line-height: 17px
+        text-overflow: ellipsis
         color: #717477
         height: fit-content
         width: fit-content
         margin: 0px 0px 0px 15px
-
-
-
-.border
-    border: 1px solid red
