+++ frontend/src/app/moduls/main/cards/recently-played-card/recently-played-card.component.ts
