+++ frontend/src/app/moduls/main/cards/top-songs-card/top-songs-card.component.html
@@ -8,4 +8,3 @@
         Ed Sheeran, Paloma Mami Maroon 5, SigalaPink, Oximer...
     </div>
 </div>
-
