+++ frontend/src/app/moduls/main/cards/top-songs-card/top-songs-card.component.sass
@@ -42,6 +42,7 @@
         font-size: 16px
         line-height: 19px
         text-align: left
+        text-overflow: ellipsis
         color: #FAFBFF
         margin: 0px 16px 5px 16px 
     .card-desc
@@ -52,8 +53,6 @@
         font-size: 14px
         line-height: 17px
         text-align: left
+        text-overflow: ellipsis
         color: #FAFBFF
         margin: 0px 16px 17px 16px
-
-
-
