+++ frontend/src/app/moduls/main/cards/top-songs-card/top-songs-card.component.spec.ts
@@ -8,9 +8,9 @@ describe('TopSongsCardComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ TopSongsCardComponent ]
+      declarations: [TopSongsCardComponent]
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
