+++ frontend/src/app/moduls/main/cards/your-mix-card/your-mix-card.component.sass
@@ -59,10 +59,6 @@
         font-weight: 500
         font-size: 16px
         line-height: 19px
+        text-overflow: ellipsis
         color: #FAFBFF 
         cursor: pointer
-
-
-
-.border
-    border: 1px solid red
