+++ frontend/src/app/moduls/main/cards/your-mix-card/your-mix-card.component.spec.ts
@@ -8,9 +8,9 @@ describe('YourMixCardComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ YourMixCardComponent ]
+      declarations: [YourMixCardComponent]
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
