+++ frontend/src/app/moduls/main/cards/your-mix-card/your-mix-card.component.ts
