+++ frontend/src/app/moduls/main/main.module.ts
@@ -1,5 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { CalmRhythmsCardComponent } from './cards/calm-rhythms-card/calm-rhythms-card.component';
+import { NewReleasesCardComponent } from './cards/new-releases-card/new-releases-card.component';
+import { RecentlyPlayedCardComponent } from './cards/recently-played-card/recently-played-card.component';
+import { TopSongsCardComponent } from './cards/top-songs-card/top-songs-card.component';
+import { YourMixCardComponent } from './cards/your-mix-card/your-mix-card.component';
 
 import { MainRoutingModule } from './main-routing.module';
 import { MainMenuComponent } from './main-menu/main-menu.component';
@@ -14,7 +19,12 @@ import { SongsComponent } from './songs/songs.component';
     MainHomeComponent,
     PlaylistComponent,
     SearchComponent,
-    SongsComponent
+    SongsComponent,
+    CalmRhythmsCardComponent,
+    NewReleasesCardComponent,
+    RecentlyPlayedCardComponent,
+    TopSongsCardComponent,
+    YourMixCardComponent
   ],
   imports: [
     CommonModule,
